Fold both ActionTestUtils implementation into test fixtures
